// Features/reports/Types/IConceptForReport.ts
export interface IConceptForReport {
    identifier: string;
    idConcept: number;
    concept: string;
    amount: number;
    conceptCode: string;
    isProfit: boolean;
    payrollName: string;
    idEmployee: number;
    employeeName: string;
    idPayrollPay: number;
    payrollPayDate: string;
    idPosition: number;
    position: string;
    idDepartment: number;
    department: string;
    idCompany: number;
}